requirements.txt
streamlit>=1.31
biopython
bcbio-gff
